moved puck.js to app/lib/
