+++ src/js/service/actionService.js
@@ -18,7 +18,6 @@ import { GridActionAudio } from '../model/GridActionAudio.js';
 import { GridActionSpeak } from '../model/GridActionSpeak.js';
 import { GridActionSpeakCustom } from '../model/GridActionSpeakCustom.js';
 import {audioUtil} from "../util/audioUtil.js";
-import {puckjs} from '../externals/puck.js';
 import {puckjsService} from './puckjsService.js';
 
 let actionService = {};
